refactor(codey-mac): extract HotkeyRecorder into a shared component
